cmake-treewide: Apply a subset of settings to in-house libiio clients
The CMake iio target is compiled with flags that enforce an extra level
of strictness. However, these flags are not currently propagated to
in-house clients such as utils/*, examples/*, and iiod. The goal is to
ensure that these flags are applied only to in-house clients without
affecting external user clients, which should not be required to
compile with these flags. To achieve this, create a CMake target of
type INTERFACE to capture these flags. Then, ensure that all in-house
clients and the CMake iio target link to this newly created interface.

Signed-off-by: Dan Nechita <dan.nechita@analog.com>