local-dmabuf: Improve checks for DMABUF interface support
To create an IIO block associated with a buffer, local_create_block()
attempts to use the DMABUF API, and if that fails it uses the MMAP API.
However, the availability check for the DMABUF API in
local_create_dmabuf() is incomplete: In recent kernels the DMABUF IOCTLs
are available, but still, the DMABUF related iio_buffer_access_funcs may
not be implemented for the buffer used in the kernel driver [1], [2]. In
this case, the returned error is propagated through local_create_block()
up to the caller and block setup fails without an attempt to use one of
the alternative interfaces.
To fix this issue, add another check for the error code set by the
DMABUF attach IOCTL and mark the DMABUF interface as unavailable.

[1]: https://elixir.bootlin.com/linux/v6.14.3/source/drivers/iio/industrialio-buffer.c#L1664
[2]: https://github.com/analogdevicesinc/libiio/issues/1278

Signed-off-by: Folker Schwesinger <dev@folker-schwesinger.de>