local: Fix disabling of IIO buffers in local_enable_buffer
Currently local_enable_buffer() can not be used to disable IIO buffers.
First, instead of propagating the value of the enable parameter to the
call to local_do_enable_buffer() a hard-coded value is used. And second,
setting both the buffer length and watermark attributes is attempted not
only in the enable case but also in the disable case for an already
active buffer. This causes local_enable_buffer() to fail early.

To fix the issue, only set buffer attributes in the enable case and
properly propagate the requested action to local_do_enable_buffer().

Signed-off-by: Folker Schwesinger <dev@folker-schwesinger.de>