Relicense specific files under MIT license #1210
As part of our ongoing efforts to enhance the flexibility and usability
of libiio, we are relicensing the following files from LGPL-2.1 to
the MIT license:

 - attr.c, attr.h
 - backend.c, block.c, buffer.c, channel.c, context.c, device.c, events.c
 - iiod/ops.h, iiod/responder.c, iiod-responder.h
 - iio-private.h
 - include/iio/iio-backend.h, include/iio/iio-debug.h, include/iio/iio.h, include/iio/iio-lock.h
 - library.c, lock-dummy.c, mask.c, scan.c, sort.c, sort.h, stream.c, task.c
 - tinyiiod/tinyiiod.c, tinyiiod/tinyiiod.h
 - utilities.c, xml.c

This change aims to better support environments where static linking is
required, such as microcontroller and baremetal/no-OS embedded systems,
and to facilitate integration with Zephyr.

The MIT license provides greater permissiveness, allowing for broader use
and integration of libiio.

All contributors to those files have provided their written consent for
this relicensing.

Signed-off-by: Michael Hennerich <michael.hennerich@analog.com>