CI: Extend the iio_toggle_feature() checking by looking at OPTIONS_LISTS
Other CMake files (such as those for iiod and the bindings) define their own options. They must also associate each option with a corresponding feature name and indicate whether the feature is enabled or disabled. This is achieved by adding the feature names to the OPTIONS_LISTS and FEATURES_LISTS.
The iio_toggle_feature() will be called for each item in the lists by the top-level cmake.
Hence the need extend the checking.

Signed-off-by: Dan Nechita <dan.nechita@analog.com>