attr: Fix context attributes getting out of sync with their values
Each time a context attribute is added, its value gets appended to a
list of values while its name gets appended to a list of names. But
since the names list gets sorted and the values list doesn't, the two
lists can become misaligned.

To fix this, after sorting, find where the newly added attribute was
inserted. Because the list gets sorted each time a new attribute is
added we know attributes positioned after the newly inserted attributes
just get shifted towards the end of the list. So we do the same for the
values.

Signed-off-by: Dan Nechita <dan.nechita@analog.com>