local: Properly initialize device fd
During device initialization iio_device_pdata is not explicitly
initialized after allocation. This may leave pdata->fd initialized to 0,
which is a valid file descriptor, corresponding to stdin of the calling
process. Later, in local_shutdown(), which is called during
iio_context_destroy(), closing of file descriptors of all IIO devices is
attempted. This effectively closes fd 0 of the calling process, making
it available for reallocation whenever a new file descriptor is
requested. This can have unintended side effects, because now, stdin of
the calling process is referring to a file that was not meant to be used
in that role.

To prevent accidental fd operations, initialize iio_device_pdata.fd to
-1, an invalid file descriptor.

Signed-off-by: Folker Schwesinger <dev@folker-schwesinger.de>