CMake: Create a separate library for wingetopt
Due to this, the compilation context of wingetopt will be different
from the one of the examples which include compiler flags that make
things more strict.
As an effect, these changes remove the warning:

getopt.c(348,22): warning C4996: 'getenv': This function or variable may
be unsafe. Consider using _dupenv_s instead. To disable deprecation, use
_CRT_SECURE_NO_WARNINGS.

Signed-off-by: Dan Nechita <dan.nechita@analog.com>