network: Translate WSA error codes to POSIX codes
It has been observed that creating a network iio context with a wrong IP
(that doesn't point to a system with iio devices) will cause a crash when
executed in a Windows environment.
This happens because:
- the windows implementation of do_create_socket() returns WSA error
- the WSA error has value -10060 which is outside of [-4095, -1]
- WSA error code is returned by network_create_context()
- iio_create_context() treats -10060 as a valid context address
- iio_create_context() calls iio_context_update_scale_offset() causing
the crash (access violation).

Fix this by translating WSA error codes to POSIX codes that are known to
be in the range of [1, 4095].
Encapsulate the translation in the Windows network implementation so
network.c doesn’t have to handle platform-specific logic. Consequently,
functions that took a WSA error and checked WSA-specific codes need to
be switched to POSIX codes.”

Signed-off-by: Dan Nechita <dan.nechita@analog.com>